requirements.txt
beautifulsoup4==4.5.3
certifi==2016.9.26
chardet==3.0.4
future==0.16.0
idna==2.7
psycopg2-binary==2.8.6
python-telegram-bot==5.3.0
requests==2.20.0
urllib3==1.24.2
